+++ lib/petal_components/table.ex
@@ -10,25 +10,38 @@ defmodule PetalComponents.Table do
 
   def table(assigns) do
     ~H"""
-    <table class={build_class([
-      "min-w-full overflow-hidden divide-y ring-1 ring-gray-200 dark:ring-0 divide-gray-200 rounded-sm table-auto dark:divide-y-0 dark:divide-gray-800 sm:rounded",
-      @class
-    ])} {@rest}>
+    <table
+      class={
+        build_class([
+          "min-w-full overflow-hidden divide-y ring-1 ring-gray-200 dark:ring-0 divide-gray-200 rounded-sm table-auto dark:divide-y-0 dark:divide-gray-800 sm:rounded",
+          @class
+        ])
+      }
+      {@rest}
+    >
       <%= render_slot(@inner_block) %>
     </table>
     """
   end
 
   attr(:class, :string, default: "", doc: "CSS class")
-  attr(:rest, :global)
+  attr(:rest, :global, include: ~w(colspan rowspan))
   slot(:inner_block, required: false)
 
   def th(assigns) do
     ~H"""
-    <th class={build_class([
-      "px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-300",
-      @class,
-    ], " ")} {@rest}>
+    <th
+      class={
+        build_class(
+          [
+            "px-6 py-3 text-xs font-medium tracking-wider text-left text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-300",
+            @class
+          ],
+          " "
+        )
+      }
+      {@rest}
+    >
       <%= render_slot(@inner_block) %>
     </th>
     """
@@ -40,25 +53,38 @@ defmodule PetalComponents.Table do
 
   def tr(assigns) do
     ~H"""
-    <tr class={build_class([
-      "border-b dark:border-gray-700 bg-white dark:bg-gray-800 last:border-none",
-      @class,
-    ])} {@rest}>
+    <tr
+      class={
+        build_class([
+          "border-b dark:border-gray-700 bg-white dark:bg-gray-800 last:border-none",
+          @class
+        ])
+      }
+      {@rest}
+    >
       <%= render_slot(@inner_block) %>
     </tr>
     """
   end
 
   attr(:class, :string, default: "", doc: "CSS class")
-  attr(:rest, :global)
+  attr(:rest, :global, include: ~w(colspan headers rowspan))
   slot(:inner_block, required: false)
 
   def td(assigns) do
     ~H"""
-    <td class={build_class([
-      "px-6 py-4 text-sm text-gray-500 dark:text-gray-400",
-      @class
-    ], " ")} {@rest}>
+    <td
+      class={
+        build_class(
+          [
+            "px-6 py-4 text-sm text-gray-500 dark:text-gray-400",
+            @class
+          ],
+          " "
+        )
+      }
+      {@rest}
+    >
       <%= render_slot(@inner_block) %>
     </td>
     """
@@ -83,8 +109,12 @@ defmodule PetalComponents.Table do
         <% end %>
 
         <div class="flex flex-col overflow-hidden">
-          <div class="overflow-hidden font-medium text-gray-900 whitespace-nowrap text-ellipsis dark:text-gray-300"><%= @label %></div>
-          <div class="overflow-hidden font-normal text-gray-500 whitespace-nowrap text-ellipsis"><%= @sub_label %></div>
+          <div class="overflow-hidden font-medium text-gray-900 whitespace-nowrap text-ellipsis dark:text-gray-300">
+            <%= @label %>
+          </div>
+          <div class="overflow-hidden font-normal text-gray-500 whitespace-nowrap text-ellipsis">
+            <%= @sub_label %>
+          </div>
         </div>
       </div>
     </div>