+++ test/petal/button_test.exs
@@ -126,7 +126,6 @@ defmodule PetalComponents.ButtonTest do
       <.icon_button to="/" link_type="button" size="xs" color="primary">
         <Heroicons.clock solid />
       </.icon_button>
-
       """)
 
     assert html =~ "<svg"
@@ -143,5 +142,17 @@ defmodule PetalComponents.ButtonTest do
 
     assert html =~ ~s{custom-attr="123"}
     assert html =~ ~s{type="button"}
+  end
+
+  test "button with icon attr" do
+    assigns = %{}
+
+    html =
+      rendered_to_string(~H"""
+      <.button icon={:home} label="Home" />
+      """)
+
+    assert html =~ "<button"
+    assert html =~ "<svg"
   end
 end