+++ test/petal/link_test.exs
@@ -93,11 +93,21 @@ defmodule PetalComponents.ATest do
            """) =~ ">Press me<"
 
     assert rendered_to_string(~H"""
-           <Link.a to="/"> Press me </Link.a>
-           """) =~ "> Press me <"
-
-    assert rendered_to_string(~H"""
            <Link.a to="/" label="Press me" />, blah
            """) =~ "</a>, blah"
+  end
+
+  test "link as a disabled button" do
+    assigns = %{}
+
+    html =
+      rendered_to_string(~H"""
+      <Link.a link_type="button" disabled to="/">
+        Press me
+      </Link.a>
+      """)
+
+    assert html =~ "Press me"
+    assert html =~ "disabled"
   end
 end